# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.CreateGeneticaDTO:
    properties:
      caracteristicas:
        type: string
      descricao:
        type: string
      nome:
        type: string
      origem:
        type: string
      plantas:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
        type: array
      tempoFloracao:
        type: integer
      tipoEspecie:
        enum:
        - cannabis
        - outra
        type: string
      tipoGenetica:
        enum:
        - sativa
        - indica
        - hibrida
        type: string
    required:
    - nome
    - origem
    - tempoFloracao
    - tipoEspecie
    - tipoGenetica
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.GeneticaResponseDTO:
    properties:
      caracteristicas:
        type: string
      descricao:
        type: string
      id:
        type: integer
      nome:
        type: string
      origem:
        type: string
      tempoFloracao:
        type: integer
      tipoEspecie:
        type: string
      tipoGenetica:
        type: string
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse:
    properties:
      data:
        items:
          type: integer
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UpdateGeneticaDTO:
    properties:
      caracteristicas:
        type: string
      descricao:
        type: string
      nome:
        type: string
      origem:
        type: string
      tempoFloracao:
        type: integer
      tipoEspecie:
        type: string
      tipoGenetica:
        type: string
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Ambiente:
    properties:
      altura:
        description: em centímetros
        type: number
      comprimento:
        description: em centímetros
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      descricao:
        type: string
      fotos:
        description: Fotos do ambiente
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Foto'
        type: array
      id:
        type: integer
      largura:
        description: em centímetros
        type: number
      microclima:
        description: Microclimas do ambiente
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Microclima'
        type: array
      nome:
        description: Nome do ambiente
        type: string
      orientacao:
        description: norte, sul, etc.
        enum:
        - norte
        - sul
        - leste
        - oeste
        type: string
      plantas:
        description: Plantas associadas a este ambiente
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
        type: array
      tempo_exposicao:
        description: em horas
        type: integer
      tipo:
        description: Tipo do ambiente (interno, externo, úmido, seco)
        enum:
        - interno
        - externo
        - humido
        - seco
        type: string
      updatedAt:
        type: string
    required:
    - altura
    - comprimento
    - largura
    - nome
    - orientacao
    - tempo_exposicao
    - tipo
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.DiarioCultivo:
    properties:
      ambientes:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Ambiente'
        type: array
      createdAt:
        type: string
      dataFim:
        type: string
      dataInicio:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nome:
        type: string
      plantas:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
        type: array
      privacidade:
        description: 'Ex: "publico", "privado"'
        type: string
      registros:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroDiario'
        type: array
      tags:
        type: string
      updatedAt:
        type: string
      usuario:
        allOf:
        - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Usuario'
        description: Relacionamentos
      usuarioID:
        type: integer
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Especie:
    enum:
    - sativa
    - indica
    - ruderalis
    type: string
    x-enum-varnames:
    - EspecieSativa
    - EspecieIndica
    - EspecieRuderalis
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Foto:
    properties:
      ambiente_id:
        description: <-- Adicione esta linha
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      descricao:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      owner_type:
        description: '"planta", "tarefa", etc.'
        type: string
      updatedAt:
        type: string
      url:
        type: string
      usuario_id:
        type: integer
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Genetica:
    properties:
      caracteristicas:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      descricao:
        type: string
      id:
        type: integer
      nome:
        description: Nome da genética
        type: string
      origem:
        description: Origem da genética
        type: string
      plantas:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
        type: array
      tempoFloracao:
        description: em dias
        type: integer
      tipoEspecie:
        description: TipoEspecie representa o tipo de espécie, como regular, feminizada,
          automatica.
        enum:
        - regular
        - feminizada
        - automatica
        type: string
      tipoGenetica:
        description: TipoGenetica representa o tipo de genética, como sativa, indica,
          ruderalis e hibrido.
        enum:
        - sativa
        - indica
        - ruderalis
        - hibrido
        type: string
      updatedAt:
        type: string
    required:
    - nome
    - origem
    - tempoFloracao
    - tipoEspecie
    - tipoGenetica
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.MeioCultivo:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      descricao:
        type: string
      id:
        type: integer
      plantas:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
        type: array
      tipo:
        description: Tipo do meio de cultivo (solo, hidroponia, aeroponia, substrato)
        enum:
        - solo
        - hidroponia
        - aeroponia
        - substrato
        type: string
      updatedAt:
        type: string
    required:
    - tipo
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Microclima:
    properties:
      ambiente_id:
        type: integer
      createdAt:
        type: string
      data_medicao:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      luminosidade:
        description: lux
        type: number
      temperatura:
        description: °C
        type: number
      umidade:
        description: '%'
        type: number
      updatedAt:
        type: string
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta:
    properties:
      ambiente:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Ambiente'
      ambiente_id:
        type: integer
      comecando_de:
        enum:
        - semente
        - clone
        type: string
      createdAt:
        type: string
      data_colheita:
        type: string
      data_plantio:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      diarios:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.DiarioCultivo'
        type: array
      especie:
        allOf:
        - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Especie'
        enum:
        - sativa
        - indica
        - ruderalis
      foto_capa:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Foto'
      foto_capa_id:
        type: integer
      genetica:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Genetica'
      genetica_id:
        type: integer
      id:
        type: integer
      meio_cultivo:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.MeioCultivo'
      meio_cultivo_id:
        type: integer
      nome:
        maxLength: 255
        minLength: 3
        type: string
      notas:
        type: string
      planta_mae:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
      planta_mae_id:
        type: integer
      registros:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroDiario'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.PlantaStatus'
        enum:
        - ativa
        - colhida
        - morta
      updatedAt:
        type: string
      usuario:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Usuario'
      usuario_id:
        type: integer
    required:
    - comecando_de
    - data_plantio
    - especie
    - nome
    - status
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.PlantaStatus:
    enum:
    - ativa
    - colhida
    - morta
    type: string
    x-enum-varnames:
    - StatusGerminating
    - StatusVegetative
    - StatusFlowering
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroDiario:
    properties:
      conteudo:
        type: string
      createdAt:
        type: string
      data:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      diario_cultivo_id:
        type: integer
      fotos:
        description: Mídia
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Foto'
        type: array
      id:
        type: integer
      planta_id:
        type: integer
      referencia_id:
        description: Campos polimórficos para associar o registro a uma entidade específica
          (opcional)
        type: integer
      referencia_tipo:
        description: '"planta", "tarefa", etc.'
        type: string
      tipo:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroTipo'
      titulo:
        type: string
      updatedAt:
        type: string
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroTipo:
    enum:
    - observacao
    - evento
    - aprendizado
    - tratamento
    - problema
    - colheita
    - crescimento
    type: string
    x-enum-varnames:
    - RegistroTipoObservacao
    - RegistroTipoEvento
    - RegistroTipoAprendizado
    - RegistroTipoTratamento
    - RegistroTipoProblema
    - RegistroTipoColheita
    - RegistroTipoCrescimento
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Tarefa:
    properties:
      ambiente_id:
        type: integer
      createdAt:
        type: string
      data_agendada:
        type: string
      data_conclusao:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      descricao:
        type: string
      diario_cultivo_id:
        type: integer
      fotos:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Foto'
        type: array
      frequencia_dias:
        description: NULL para tarefas únicas
        type: integer
      id:
        type: integer
      planta_id:
        description: NULL para tarefas gerais
        type: integer
      prioridade:
        description: baixa, média, alta
        type: string
      recorrente:
        type: boolean
      status:
        description: pendente, concluída, atrasada
        type: string
      tipo:
        description: regar, adubar, podar, etc.
        type: string
      updatedAt:
        type: string
      usuario_id:
        type: integer
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Usuario:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
      plantas:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Planta'
        type: array
      preferencias:
        description: Configurações em JSON
        items:
          type: integer
        type: array
      tarefas:
        items:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.Tarefa'
        type: array
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Cultivo app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cultivo API
  version: "1.0"
paths:
...
  /api/v1/geneticas:
    get:
      description: Retorna uma lista paginada de todas as genéticas cadastradas
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limite de itens por página (padrão: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.GeneticaResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista todas as genéticas com paginação
      tags:
      - genetica
    post:
      consumes:
      - application/json
      description: Cria uma nova genética com os dados fornecidos
      parameters:
      - description: Dados da Genética
        in: body
        name: genetica
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.CreateGeneticaDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.GeneticaResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria uma nova genética
      tags:
      - genetica
  /api/v1/geneticas/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta uma genética existente
      parameters:
      - description: ID da Genética
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta uma genética
      tags:
      - genetica
    get:
      consumes:
      - application/json
      description: Retorna informações detalhadas de uma genética
      parameters:
      - description: Genetica ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.GeneticaResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca uma genética por ID
      tags:
      - genetica
    put:
      consumes:
      - application/json
      description: Atualiza uma genética existente com os dados fornecidos
      parameters:
      - description: ID da Genética
        in: path
        name: id
        required: true
        type: integer
      - description: Dados da Genética para atualização
        in: body
        name: genetica
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UpdateGeneticaDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.GeneticaResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza uma genética
      tags:
      - genetica
...
swagger: "2.0"